+++ R/utils_register_parallel.R
@@ -4,3 +4,13 @@ register_parallel <- function(ncores) {
   doParallel::registerDoParallel(cl)
   return(cl)
 }
+
+
+.options_to_cluster <- function(keyword) {
+  relevant_option_pos <- grep(
+    pattern = keyword,
+    x = names(options()),
+    value = FALSE
+  )
+  return(options()[relevant_option_pos])
+}