+++ R/zzz.R
@@ -4,11 +4,12 @@ NULL
 
 mlexperiments_default_options <- list(
   mlexperiments.learner = c(
-    MLSurvGlmnetCox$classname # = "MLSurvGlmnetCox"
+    LearnerSurvGlmnetCox$classname # = "LearnerSurvGlmnetCox"
   ),
-  mlexperiments.xgb.nrounds = 5000L,
-  mlexperiments.xgb.early_stopping_rounds = 500L,
-  mlexperiments.bayesian.max_init = 50L
+  mlexperiments.optim.xgb.nrounds = 5000L,
+  mlexperiments.optim.xgb.early_stopping_rounds = 500L,
+  mlexperiments.bayesian.max_init = 50L,
+  mlexperiments.xgb.print_every_n = 50L
 )